Handle 404 exception and add DTO layer

+++ konpeito-backend/src/main/java/com/tertal/konpeito/repository/ApplicationRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.tertal.konpeito.entity.Application;
 
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application, Integer> {
+public interface ApplicationRepository extends JpaRepository<Application, Long> {
 
     @Query("""
     SELECT a
